% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_fstat.R, R/write_genetic.R
\name{read_genetic_data}
\alias{read_fstat}
\alias{read_genetic_data}
\alias{write_genetic}
\title{Read data genetic data}
\usage{
read_fstat(fn)

write_genetic(x, fn, loc_sep = ":", sep = ",")
}
\arguments{
\item{fn}{Character string, the file name.}

\item{x}{A three dimensional array containing the genetic data.}

\item{loc_sep}{Character, the separator to devide alleles.}

\item{sep}{Character, the field separator.}
}
\value{
A three dimensional array to be used with the \code{fpga} package. Population names a stored as row names.
}
\description{
The function \code{read_fstat} reads data stored in the Fstat file format (http://heidi.chnebu.ch/doku.php?id=fstat). Typicall Fstat files have the extension \code{*.dat}.
}

