% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/het.R, R/n_alls.R, R/n_eff_alls.R, R/private_alleles.R
\name{het_obs}
\alias{het_exp}
\alias{het_obs}
\alias{metrics}
\alias{n_alls}
\alias{n_eff_alls}
\alias{nprivate_alleles}
\title{Population genetic metrics}
\usage{
het_obs(x, stratum)

het_exp(x, stratum)

n_alls(x, stratum)

n_eff_alls(x, stratum)

nprivate_alleles(x, stratum)
}
\arguments{
\item{x}{Three dimensional array with individuals, loci and alleles in the first, second and third dimension.}

\item{stratum}{Vector of length \code{nrow(x)} indicating population membership.}
}
\description{
Functions to calculate population genetic metrics. For all functions \code{stratum} splits the data set into different populations. If \code{stratum} is missing, one population is assumed. \code{n_alls} returns the number of alleles per population.
}

