% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/freq.R, R/het.R, R/n_alls.R, R/n_eff_alls.R, R/private_alleles.R, R/rarify.R
\name{freq}
\alias{freq}
\alias{het_exp}
\alias{het_obs}
\alias{metrics}
\alias{n_alls}
\alias{n_eff_alls}
\alias{nprivate_alleles}
\alias{rarify}
\title{Population genetic metrics}
\usage{
freq(x, stratum)

het_obs(x, stratum)

het_exp(x, stratum)

n_alls(x, stratum)

n_eff_alls(x, stratum)

nprivate_alleles(x, stratum)

rarify(x, FUN, stratum, nperm = 100)
}
\arguments{
\item{x}{Three dimensional array with individuals, loci and alleles in the first, second and third dimension.}

\item{stratum}{Vector of length \code{nrow(x)} indicating population membership.}

\item{FUN}{Function to rarify.}

\item{nperm}{Number of permutations.}
}
\description{
Functions to calculate population genetic metrics. For all functions \code{stratum} splits the data set into different populations. If \code{stratum} is missing, one population is assumed. \code{n_alls} returns the number of alleles per population. \code{rarify} rarifies a population genetic metric.
}

