% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dest.R, R/gst.R, R/gst_p.R, R/pw_gstructure.R
\name{dest}
\alias{dest}
\alias{gst}
\alias{gst_p}
\alias{pw_gstructure}
\alias{structure}
\title{Population genetic structure}
\usage{
dest(x, stratum, size.correct = FALSE)

gst(x, stratum, size.correct = FALSE)

gst_p(x, stratum, size.correct = FALSE)

pw_gstructure(x, stratum, size.correct = FALSE, what = c("gst", "gst_p",
  "dest"))
}
\arguments{
\item{x}{Three dimensional array with individuals, loci and alleles in the first, second and third dimension.}

\item{stratum}{Vector of length \code{nrow(x)} indicating population membership.}

\item{size.correct}{Logical value indicating whether or not to correct for sample size.}
}
\description{
Functions to population genetic differences between populations.
}

